+++ Client/Rendering/g3d/include/G3D/Color3.h
@@ -136,11 +136,11 @@ public:
 
     }
 
-    inline double sum() const {
+    inline float sum() const {
         return r + g + b;
     }
 
-    inline double average() const {
+    inline float average() const {
         return sum() / 3.0;
     }
